+++ packages/qdadm/src/kernel/index.js
@@ -16,3 +16,8 @@ export {
   EventRouter,
   createEventRouter,
 } from './EventRouter.js'
+export {
+  SSEBridge,
+  createSSEBridge,
+  SSE_SIGNALS,
+} from './SSEBridge.js'